+++ examples/.experimental/next-app-dir-react-query/src/app/MyPage.tsx
@@ -1,0 +1,42 @@
+'use client';
+
+import { useQuery } from '@tanstack/react-query';
+import { Suspense } from 'react';
+import { MyComponent } from './MyComponent';
+
+export function MyPage() {
+  return (
+    <Suspense>
+      <Suspense fallback={<div>100</div>}>
+        <MyComponent wait={100} />
+      </Suspense>
+      <Suspense fallback={<div>200</div>}>
+        <MyComponent wait={200} />
+      </Suspense>
+      <Suspense fallback={<div>300</div>}>
+        <MyComponent wait={300} />
+      </Suspense>
+      <Suspense fallback={<div>400</div>}>
+        <MyComponent wait={400} />
+      </Suspense>
+      <Suspense fallback={<div>500</div>}>
+        <MyComponent wait={500} />
+      </Suspense>
+      <Suspense fallback={<div>600</div>}>
+        <MyComponent wait={600} />
+      </Suspense>
+      <Suspense fallback={<div>700</div>}>
+        <MyComponent wait={700} />
+      </Suspense>
+      <Suspense fallback={<div>800</div>}>
+        <MyComponent wait={800} />
+      </Suspense>
+      <Suspense fallback={<div>900</div>}>
+        <MyComponent wait={900} />
+      </Suspense>
+      <Suspense fallback={<div>1000</div>}>
+        <MyComponent wait={1000} />
+      </Suspense>
+    </Suspense>
+  );
+}
