+++ examples/.experimental/next-app-dir-react-query/src/app/page.tsx
@@ -1,4 +1,4 @@
-import { MyPage } from './MyComponent';
+import { MyPage } from './MyPage';
 
 export default function Page() {
   return (
